Elijah's father trying to help turn it in
Hello, this is Elijah's father trying to help him turn in his assignment as it seems that there is a software glitch that won't allow him to submit it? I am trying to help him problem solve to see if there is any error on our side that we can work past. I hope to solve this so it can finally be submitted!